boom: add --check-bootloader
Add a way to run the bootloader check without having to run an
entry create command.

This is a bit of a kludge atm: the logic for the object/verb style
requires at least one argument, so the option needs another valid
command word on the to pass the arg checks:

 # boom --check-bootloader
 boom: error: too few arguments

  # boom create --check-bootloader
  WARNING - Boom grub2 script missing from '/etc/grub.d'
  WARNING - Boom configuration not found in grub.cfg
  WARNING - Run 'grub2-mkconfig > /boot/grub2/grub.cfg' to enable

This could be fixed by changing the argument checking (without
affecting validation of object/verb commands), or by adding a new
'system' object type with a different command table.

Not needed immediately so parking on a branch for now.