command: add add_opts (--add-option) and del_opts (--del-option)
Add new `add_opts` and `del_opts` kwargs to the `create_entry`,
and `clone_entry` API functions. These mirror the `add_opts` and
`del_opts` arguments to the `BootParams` initialiser and allow
users of the API to specify kernel options to be added or deleted.

The options are used by the command handlers for create and clone
to allow the user to specify options to be added with --add-opts
and removed with --del-opts. Each option takes a space-separated
list of options with or without values (using name=value notation).

Related: #16

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>