command: add support for automatic titles
Allow the boom.command interfaces to pass down an empty title
when creating entries. If the corresponding profiles define a
title template key then this will be used to generate the title
for the new entry. Passing an empty title where the attached
profiles do not define a title key results in error.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>