bootloader: make format key substitution table-driven
Replace open-coded key substitution with a general table-driven
approach.

A table is defined that contains descriptions of the formatting
and data dependencies required for each substitutable key.

Each entry in the format_key_specs table specifies a list of
possible key substitutions to perform for the named key. Each
entry of the key_spec list contains a dictionary containing
one or more attribute sources or predicates. An optional value
format allows re-formatting of the attribute value before
substitution takes place (for e.g. to add 'name=' markup).

A key substitution is evaluated if at least one of the listed
attribute sources is defined, and if all defined predicates
evaluate to True. A predicate must be a Python callable
accepting no arguments and returning a boolean. A key_spec
may also specify an explicit list of needed objects, "bp",
or "osp", that must exist to evaluate predicates.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>