profiles: share more methods between OsProfile and HostProfile
Rename the underlying dictionary in HostProfile back to the same
name as OsProfile: _profile_data. This allows more methods and
attributes to be shared with the parent class:

  * __len__(), __getitem__(), keys(), values(), items(), _dirty()
    __from_file()

A pair of new members are added to the base OsProfile class:

  _required_keys, and _identity_key

These store the list of required keys for the profile type, and
the name of the key holding the profile identifier repectively.

With additional abstraction of profile details further methods
can be made shared: write_profile(), delete_profile() and the
dictionary based loader __from_data().

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>