boom: export BoomLogger class
The BoomLogger class is not very useful to client classes (though
it could be used to create an addtional subclass to perform more
log filtering in an application using boom). Its presence adds a
little clutter but isn't a real problem.

It does need to be tested by the unit suite though, so add it to
the __all__ member for the module.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>