tests: fix OsProfile test lifecycles and uniqueness
Now that the OsProfile duplicate test is handled directly in the
class initialiser (rather than in the boom.command wrapper), make
sure that all OsProfiles used in tests are either written out to
disk as expected, or, if they are transient, that they are deleted
at the end of each test case.

Additionally, ensure that OsProfiles are unique when that is a
requirement for a test and that all test entries using a test
profile have a valid (non-null) profile identifier set.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>