command: factor out add_opts/del_opts merging for clone/edit
Both clone and edit need to be able to take a command-line add_opts
or del_opts string, and merge it with any existing add or del opts
list contained in the BootParams of the BootEntry being edited or
cloned. Factor this out into a separate _merge_add_del_opts()
function and use it in the two API calls.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>