boom.bootloader: mark foreign BLS entries as read-only
If a boot entry is loaded that does not conform to the expected
file name pattern for Boom boot entries, mark it as read-only.

This prevents boom from unexpectedly modifying boot entries that
were created by other tools, or that are under distribution package
management control.

Users of the API can disable this by setting be.read_only=False
after loading: if the entry is subsequently written it will be
given a new boom-compliant path name and be written as a native
boom entry.

Users of the CLI can simply clone foreign entries rather than
editing them.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>